refacto/clean for agency right to review
